+++ functions/util/admin.js
@@ -1,0 +1,6 @@
+const admin = require('firebase-admin');
+
+admin.initializeApp();
+const db = admin.firestore();
+
+module.exports = {admin, db};
