+++ functions/util/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  apiKey: "AIzaSyAt9i5XiZyom2mXsOSpLfd_0ILgDZ6uLhM",
+  authDomain: "socialdemospa.firebaseapp.com",
+  databaseURL: "https://socialdemospa.firebaseio.com",
+  projectId: "socialdemospa",
+  storageBucket: "socialdemospa.appspot.com",
+  messagingSenderId: "456078504994",
+  appId: "1:456078504994:web:236a86e05020e2b717867b",
+  measurementId: "G-17M8KBTB0B"
+};
