+++ functions/util/firebaseAuth.js
@@ -1,0 +1,29 @@
+const {admin} = require('./admin');
+
+exports.firebaseAuth = (request, response, next) => {
+  let idToken;
+  if(request.headers.authorization && request.headers.authorization.startsWith('Bearer ')){
+    idToken = request.headers.authorization.split('Bearer ')[1];
+  } else {
+    console.error('No token found.')
+    return response.status(403).json({ error: 'Unauthorized'});
+  }
+
+  return admin.auth().verifyIdToken(idToken)
+    .then(decodedToken => {
+      request.user = decodedToken;
+      return db.collection('users')
+        .where('userId', '==', request.user.uid)
+        .limit(1)
+        .get();
+    })
+    .then( data => {
+      request.user.userName = data.docs[0]
+        .data().userName
+        return next()
+    })
+    .catch((error) => {
+      console.error('verify token error', error)
+      return response.status(403).json(error);
+    });
+};
